refactor(reducers): Move folders into 'users' namespace.
