+++ src/app/users/models/users.ts
@@ -4,19 +4,17 @@ import { Observable } from 'rxjs/Observable';
 import { BehaviorSubject } from 'rxjs/subject/BehaviorSubject';
 import { normalize, arrayOf, Schema } from 'normalizr';
 import { List, Map, Record, fromJS } from 'immutable';
-import {ApiService} from '../modules/api';
+import { ApiService } from '../services/api';
 
-import {
-  LOADING_USERS,
-  LOADED_USERS,
-  LOADING_USER,
-  LOADED_USER,
-  ADDING_USER,
-  ADDED_USER,
-  DELETING_USER,
-  DELETED_USER,
-  PATCHED_USER
-} from '../reducers/users';
+export const LOADING_USERS = 'LOADING_USERS';
+export const LOADED_USERS = 'LOADED_USERS';
+export const LOADING_USER = 'LOADING_USER';
+export const LOADED_USER = 'LOADED_USER';
+export const ADDING_USER = 'ADDING_USER';
+export const ADDED_USER = 'ADDED_USER';
+export const DELETING_USER = 'DELETING_USER';
+export const DELETED_USER = 'DELETED_USER';
+export const PATCHED_USER = 'PATCHED_USER';
 
 const PATCH_USER = 'PATCH_USER';
 const DELETE_USER = 'DELETE_USER';
